fix: thresholds of check rules are floating point numbers (#45)

+++ internal/types/check_rule.go
@@ -74,8 +74,8 @@ type Dash0CheckRule struct {
 }
 
 type Dash0CheckRuleThresholds struct {
-	Degraded int `json:"degraded"`
-	Failed   int `json:"failed"`
+	Degraded float64 `json:"degraded"`
+	Failed   float64 `json:"failed"`
 }
 
 type PrometheusRules struct {
